+++ src/schema/2.0/example-object.ts
@@ -1,0 +1,6 @@
+import { Dictionary } from '../../utils/types';
+import * as t from 'io-ts';
+
+export interface ExampleObject extends Dictionary<string> {}
+
+export const ExampleObject = t.record(t.string, t.string, 'ExampleObject');
