+++ src/schema/2.0/headers-object.ts
@@ -1,0 +1,7 @@
+import { Dictionary } from '../../utils/types';
+import { HeaderObject } from './header-object';
+import * as t from 'io-ts';
+
+export interface HeadersObject extends Dictionary<HeaderObject> {}
+
+export const HeadersObject = t.record(t.string, HeaderObject, 'HeadersObject');
