+++ src/schema/2.0/license-object.ts
@@ -1,0 +1,16 @@
+import { Option } from 'fp-ts/lib/Option';
+import * as t from 'io-ts';
+import { stringOption } from '../../utils/io-ts';
+
+export interface LicenseObject {
+	readonly name: string;
+	readonly url: Option<string>;
+}
+
+export const LicenseObject = t.type(
+	{
+		name: t.string,
+		url: stringOption,
+	},
+	'LicenseObject',
+);
