+++ src/schema/2.0/parameters-definitions-object.ts
@@ -1,0 +1,7 @@
+import { Dictionary } from '../../utils/types';
+import { ParameterObject } from './parameter-object/parameter-object';
+import * as t from 'io-ts';
+
+export interface ParametersDefinitionsObject extends Dictionary<ParameterObject> {}
+
+export const ParametersDefinitionsObject = t.record(t.string, ParameterObject, 'ParametersDefinitionsObject');
