+++ src/schema/2.0/reference-object.ts
@@ -1,0 +1,12 @@
+import * as t from 'io-ts';
+
+export interface ReferenceObject {
+	readonly $ref: string;
+}
+
+export const ReferenceObject = t.type(
+	{
+		$ref: t.string,
+	},
+	'ReferenceObject',
+);
