+++ src/schema/2.0/responses-definitions-object.ts
@@ -1,0 +1,7 @@
+import { Dictionary } from '../../utils/types';
+import { ResponseObject } from './response-object';
+import * as t from 'io-ts';
+
+export interface ResponsesDefinitionsObject extends Dictionary<ResponseObject> {}
+
+export const ResponsesDefinitionsObject = t.record(t.string, ResponseObject, 'ResponsesDefinitionsObject');
