+++ src/schema/2.0/responses-object.ts
@@ -1,0 +1,7 @@
+import { Dictionary } from '../../utils/types';
+import { ResponseObject } from './response-object';
+import * as t from 'io-ts';
+
+export interface ResponsesObject extends Dictionary<ResponseObject> {}
+
+export const ResponsesObject = t.record(t.string, ResponseObject, 'ResponsesObject');
