+++ src/schema/2.0/schema-object/boolean-property-schema-object.ts
@@ -1,0 +1,12 @@
+import * as t from 'io-ts';
+
+export interface BooleanPropertySchemaObject {
+	readonly type: 'boolean';
+}
+
+export const BooleanPropertySchemaObject = t.type(
+	{
+		type: t.literal('boolean'),
+	},
+	'BooleanPropertySchemaObject',
+);
