+++ src/schema/2.0/schema-object/number-property-schema-object.ts
@@ -1,0 +1,16 @@
+import * as t from 'io-ts';
+import { stringOption } from '../../../utils/io-ts';
+import { Option } from 'fp-ts/lib/Option';
+
+export interface NumberPropertySchemaObject {
+	readonly type: 'number';
+	readonly format: Option<string>;
+}
+
+export const NumberPropertySchemaObject = t.type(
+	{
+		type: t.literal('number'),
+		format: stringOption,
+	},
+	'NumberPropertySchemaObject',
+);
