+++ src/schema/2.0/scopes-object.ts
@@ -1,0 +1,6 @@
+import { Dictionary } from '../../utils/types';
+import * as t from 'io-ts';
+
+export interface ScopesObject extends Dictionary<string> {}
+
+export const ScopesObject = t.record(t.string, t.string, 'ScopesObject');
