+++ src/schema/2.0/security-scheme-object/base-security-scheme-object.ts
@@ -1,0 +1,10 @@
+import { stringOption } from '../../../utils/io-ts';
+import { Option } from 'fp-ts/lib/Option';
+
+export interface BaseSecuritySchemeObjectProps {
+	readonly description: Option<string>;
+}
+
+export const BaseSecuritySchemeObjectProps = {
+	description: stringOption,
+};
