+++ src/schema/2.0/security-scheme-object/basic-security-scheme-object.ts
@@ -1,0 +1,14 @@
+import * as t from 'io-ts';
+import { BaseSecuritySchemeObjectProps } from './base-security-scheme-object';
+
+export interface BasicSecuritySchemeObject extends BaseSecuritySchemeObjectProps {
+	readonly type: 'basic';
+}
+
+export const BasicSecuritySchemeObject = t.type(
+	{
+		...BaseSecuritySchemeObjectProps,
+		type: t.literal('basic'),
+	},
+	'BasicSecuritySchemeObject',
+);
