+++ src/utils.ts
@@ -1,23 +1,21 @@
 import { array, uniq, flatten, last } from 'fp-ts/lib/Array';
-import {
-	PathItemObject,
-	OperationObject,
-	PathsObject,
-	Dictionary,
-	ParameterObject,
-	PathParameterObject,
-	ReferenceObject,
-	BodyParameterObject,
-	SwaggerObject,
-	ParametersDefinitionsObject,
-	QueryParameterObject,
-} from './schema/2.0/swagger';
 import { constant, Endomorphism, identity, tuple } from 'fp-ts/lib/function';
 import { getStructEq, eqString } from 'fp-ts/lib/Eq';
 import { FSEntity } from './fs';
 import { camelize } from '@devexperts/utils/dist/string/string';
 import { alt, map, mapNullable, option, Option, some, chain, getOrElse } from 'fp-ts/lib/Option';
 import { pipe } from 'fp-ts/lib/pipeable';
+import { SwaggerObject } from './schema/2.0/swagger-object';
+import { Dictionary } from './utils/types';
+import { ReferenceObject } from './schema/2.0/reference-object';
+import { QueryParameterObject } from './schema/2.0/parameter-object/query-parameter-object';
+import { PathParameterObject } from './schema/2.0/parameter-object/path-parameter-object';
+import { BodyParameterObject } from './schema/2.0/parameter-object/body-parameter-object';
+import { ParameterObject } from './schema/2.0/parameter-object/parameter-object';
+import { OperationObject } from './schema/2.0/operation-object';
+import { PathItemObject } from './schema/2.0/path-item-object';
+import { PathsObject } from './schema/2.0/paths-object';
+import { ParametersDefinitionsObject } from './schema/2.0/parameters-definitions-object';
 
 export interface Serializer {
 	(name: string, schema: SwaggerObject): FSEntity;
