+++ src/utils/types.ts
@@ -1,0 +1,1 @@
+export interface Dictionary<A> extends Readonly<Record<string, A>> {}
